Add Hot buttered rum cocktail + some other updates (#535)
* Update a furlong too late

Slightly updated wording. "the" implies it's in ingredient list, but it's not. "a" better implies that you can add this garnish but it's not specifically called out as an ingredient.

* Update Absolut Limousine recipe

Updated instructions for Absolut Limousine to refer to ingredients as listed, and include the tonic water ingredient in the instructions.

* Add Hot Buttered Rum

Added new recipe Hot Buttered Rum.