Update to Aquafaba Gin Fizz recipe
Bumped up the soda water amount. Most recipes I found called for filling the rest of the glass, and one called for 30ml/1oz. Have gone with 30ml for now as this seems to be the minimum you'd want.

Adjusted instructions to refer to simple syrup like the ingredient list instead of sugar syrup.