Updated ingredients/method of black margarita
I have rewritten the Black Margarita recipe. Added maple syrup and cointreau to ingredients list as these are critical parts but are not listed in ingredient list. Removed simple syrup from ingredients as this is not referenced in instructions.

Rewrote recipe to be more in line with other recipes, and make some parts clearer.