Align recipes for cocktails starting with B + more (#564)
* Some instruction updates

Punctuation added to B-52

Changed wording of Balmoral to match other cocktails.

Removed part about service banana milkshake in wine glass, since this doesn't match the picture.

* Update barracuda instructions

Updated instructions to align with other recipes.

Referred to "Prosecco" rather than "sparkling wine" as ingredients list uses "Prosecco", but left "sparkling wine" reference at end.

* Update bees knees

Updated bees knees instructions to separate steps a little better

* Update Belgian blue instructions

Aligned language with other cocktails

* Update Bellini recipe

Minor changes

* Update Bermuda highball

Minor changes to wording of recipe to align with others.

* Update allegheny recipe

Updated allegheny recipe to align with others. Doesn't start with B but I seem to have missed this last time.

* Update between he sheets

Updated wording of recipe to align with other cocktails

* Updated ingredients/method of black margarita

I have rewritten the Black Margarita recipe. Added maple syrup and cointreau to ingredients list as these are critical parts but are not listed in ingredient list. Removed simple syrup from ingredients as this is not referenced in instructions.

Rewrote recipe to be more in line with other recipes, and make some parts clearer.

* Update blackthorn wording

Minor update to wording to show you can garnish with a twist of lemon peel. Original wording implies it's listed in ingredients when it is not.

* Update bloody mary instructions

Adjusted wording of instructions to align with other cocktails.

* Major update to bloody punch

The bloody punch recipe is hard to follow and intended for a punch bowl. It is hard to align the recipe with the single person ingredients without losing the simplicity. I've updated the bloody punch recipe to use the original large sizes intended for a punch bowl, and adjusted the recipe to align with others. The option to choose multiple servings is therefore to choose to create larger bowls of punch.

* Update blue-lagoon instructions

Minor change to wording to align with other recipes

* Minor update to bluebird instructions

Minor update to have garnish as extras instead of implying it was in ingredients list

* Update bombay cassis

Updated recipe to remove branding and align with other recipes.

* Update boulevardier instructions

Updated to align with other recipes

* Update bounty hunter

Align recipe with other recipes.

* Update boxcar recipe

Updated to align with other recipes.

* Remove cream soda cocktail

This Cream Soda cocktail is basically the same as the Happy Skipper cocktail, and even has the same picture. I have removed this as the Cream Soda cocktail appears to actually be something else.

* Update bramble recipe

Minor changes to Bramble to align with other recipes